requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.3
dj-database-url==1.0.0
Django==4.1.3
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
